chore: Introduce a Taskfile to act as our task-runner
Internally, Taskfile has been replacing our use of Makefile and PyInvoke.

It allows for easy discovery of tasks; and is a language-neutral task runner.

The taskfile is structured in a way that allows for standalone e2e testing, while
being injectable so that our monorepo can run a full e2e test.